requirements.txt
ansible==9.4.0
ansible-core==2.16.6
ansible-runner==2.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.5
docutils==0.21.1
idna==3.7
Jinja2==3.1.3
lockfile==0.12.2
MarkupSafe==2.1.5
packaging==24.0
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
pyspnego==0.10.2
python-daemon==3.0.1
pywinrm==0.4.3
PyYAML==6.0.1
requests==2.31.0
requests-ntlm==1.2.0
resolvelib==1.0.1
six==1.16.0
urllib3==2.2.1
xmltodict==0.13.0
